+++ src/app/(content-layout)/[groupId]/(groupId)/_[groupId]/Tasklists/index.tsx
@@ -1,18 +1,12 @@
 'use client';
-import { useState, useOptimistic, useTransition } from 'react';
+import useTasklists from '@/app/(content-layout)/[groupId]/(groupId)/_[groupId]/Tasklists/useTasklists';
 import TasklistItem from '@/app/(content-layout)/[groupId]/(groupId)/_[groupId]/Tasklists/TasklistItem';
 import TasklistCreateModal from '@/app/(content-layout)/[groupId]/(groupId)/_[groupId]/Tasklists/TasklistCreateModal';
 import TasklistUpdateModal from '@/app/(content-layout)/[groupId]/(groupId)/_[groupId]/Tasklists/TasklistUpdateModal';
 import TasklistDeleteModal from '@/app/(content-layout)/[groupId]/(groupId)/_[groupId]/Tasklists/TasklistDeleteModal';
-import {
-  createTasklistAction,
-  updateTasklistAction,
-  deleteTasklistAction,
-} from '@/app/(content-layout)/[groupId]/(groupId)/_[groupId]/Tasklists/actions';
 import { ModalTrigger } from '@/components/common/modal';
 import { Group } from '@/types/group';
 import { Tasklist } from '@/types/tasklist';
-import useModalContext from '@/components/common/modal/core/useModalContext';
 
 type TasklistsProps = {
   groupId: Group['id'];
@@ -20,99 +14,18 @@ type TasklistsProps = {
 };
 
 export default function Tasklists({ groupId, tasklists }: TasklistsProps) {
-  const [optimisticTasklists, setOptimisticTasklists] = useOptimistic(
-    tasklists,
-    (
-      currentTasklists: Tasklist[],
-      action:
-        | { type: 'create'; newTasklist: Tasklist }
-        | { type: 'update'; selectedTasklist: Tasklist; newName: Tasklist['name'] }
-        | { type: 'delete'; selectedTasklist: Tasklist }
-        | { type: 'rollback' }
-    ) => {
-      if (action.type === 'create') {
-        return [...currentTasklists, action.newTasklist];
-      } else if (action.type === 'update') {
-        return currentTasklists.map((tasklist) =>
-          tasklist.id === action.selectedTasklist.id
-            ? { ...tasklist, name: action.newName }
-            : tasklist
-        );
-      } else if (action.type === 'delete') {
-        return currentTasklists.filter((tasklist) => tasklist.id !== action.selectedTasklist.id);
-      } else if (action.type === 'rollback') {
-        return currentTasklists;
-      }
-      return currentTasklists;
-    }
-  );
-  const [selectedTasklist, setSelectedTasklist] = useState<Tasklist | null>(null);
-  const [isCreationLoading, startCreationTransition] = useTransition();
-  const [isUpdateLoading, startUpdateTransition] = useTransition();
-  const [isDeletionLoading, startDeletionTransition] = useTransition();
-  const [transitionError, setTransitionError] = useState<{ message: string; id: string } | null>(
-    null
-  );
-
-  const createTasklist = async (name: string) => {
-    startCreationTransition(async () => {
-      const newTasklist = {
-        id: -1,
-        name: name,
-        createdAt: '',
-        updatedAt: '',
-        groupId: groupId,
-        displayIndex: tasklists[tasklists.length - 1].displayIndex + 1,
-        tasks: [],
-      };
-      setTransitionError(null);
-      setOptimisticTasklists({ type: 'create', newTasklist });
-
-      const result = await createTasklistAction(groupId, name);
-
-      if (!result.success) {
-        setOptimisticTasklists({ type: 'rollback' });
-        setTransitionError({
-          message: result.message,
-          id: Date.now().toString(),
-        });
-      }
-    });
-  };
-
-  const updateTasklist = async (selectedTasklist: Tasklist, newName: string) => {
-    startUpdateTransition(async () => {
-      setTransitionError(null);
-      setOptimisticTasklists({ type: 'update', selectedTasklist, newName });
-
-      const result = await updateTasklistAction(groupId, selectedTasklist.id, newName);
-
-      if (!result.success) {
-        setOptimisticTasklists({ type: 'rollback' });
-        setTransitionError({
-          message: result.message,
-          id: Date.now().toString(),
-        });
-      }
-    });
-  };
-
-  const deleteTasklist = async (selectedTasklist: Tasklist) => {
-    startDeletionTransition(async () => {
-      setTransitionError(null);
-      setOptimisticTasklists({ type: 'delete', selectedTasklist });
-
-      const result = await deleteTasklistAction(groupId, selectedTasklist.id);
-
-      if (!result.success) {
-        setOptimisticTasklists({ type: 'rollback' });
-        setTransitionError({
-          message: result.message,
-          id: Date.now().toString(),
-        });
-      }
-    });
-  };
+  const {
+    optimisticTasklists,
+    selectedTasklist,
+    setSelectedTasklist,
+    isCreateLoading,
+    isUpdateLoading,
+    isDeleteLoading,
+    transitionError,
+    createTasklist,
+    updateTasklist,
+    deleteTasklist,
+  } = useTasklists(groupId, tasklists);
 
   const tasklistCreateModalId = `tasklistCreate-${groupId}`;
   const tasklistUpdateModalId = selectedTasklist ? `tasklistUpdate-${selectedTasklist.id}` : '';
@@ -144,8 +57,8 @@ export default function Tasklists({ groupId, tasklists }: TasklistsProps) {
       <TasklistCreateModal
         modalId={tasklistCreateModalId}
         createTasklist={createTasklist}
-        isCreationLoading={isCreationLoading}
-        errorOnCreate={transitionError}
+        isLoading={isCreateLoading}
+        error={transitionError}
       />
 
       {selectedTasklist && (
@@ -153,8 +66,8 @@ export default function Tasklists({ groupId, tasklists }: TasklistsProps) {
           tasklist={selectedTasklist}
           modalId={tasklistUpdateModalId}
           updateTasklist={updateTasklist}
-          isUpdateLoading={isUpdateLoading}
-          errorOnUpdate={transitionError}
+          isLoading={isUpdateLoading}
+          error={transitionError}
         />
       )}
 
@@ -163,8 +76,8 @@ export default function Tasklists({ groupId, tasklists }: TasklistsProps) {
           tasklist={selectedTasklist}
           modalId={tasklistDeleteModalId}
           deleteTasklist={deleteTasklist}
-          isDeletionLoading={isDeletionLoading}
-          errorOnDelete={transitionError}
+          isLoading={isDeleteLoading}
+          error={transitionError}
         />
       )}
     </>
